backend/requirements.txt
Flask>=3.0.0,<4
gunicorn>=22.0.0,<23
psycopg2-binary>=2.9.0,<3

